--- a/websocietysimulator/tools/cache_interaction_tool.py
+++ b/websocietysimulator/tools/cache_interaction_tool.py
@@ -1,8 +1,9 @@
 import logging
 import os
 import json
+import lmdb
 from typing import Optional, Dict, List, Iterator
-from cachetools import LRUCache
+from tqdm import tqdm
 
 logger = logging.getLogger("websocietysimulator")
 
@@ -12,15 +13,69 @@ class CacheInteractionTool:
         Initialize the tool with the dataset directory.
         Args:
             data_dir: Path to the directory containing Yelp dataset files.
-            cache_size: Maximum number of items to keep in each cache.
+            map_size: Maximum size of the database in bytes.
         """
         logger.info(f"Initializing InteractionTool with data directory: {data_dir}")
         self.data_dir = data_dir
-        self.user_cache = LRUCache(maxsize=cache_size)
-        self.item_cache = LRUCache(maxsize=cache_size)
-        self.review_cache = LRUCache(maxsize=cache_size)
-        self.item_reviews_cache = LRUCache(maxsize=cache_size)
-        self.user_reviews_cache = LRUCache(maxsize=cache_size)
+
+        # Create LMDB environments
+        self.env_dir = os.path.join(data_dir, "lmdb_cache")
+        os.makedirs(self.env_dir, exist_ok=True)
+
+        self.user_env = lmdb.open(os.path.join(self.env_dir, "users"), map_size=5 * 1024 * 1024 * 1024)
+        self.item_env = lmdb.open(os.path.join(self.env_dir, "items"), map_size=5 * 1024 * 1024 * 1024)
+        self.review_env = lmdb.open(os.path.join(self.env_dir, "reviews"), map_size=50 * 1024 * 1024 * 1024)
+
+        # Initialize the database if empty
+        self._initialize_db()
+
+    def _initialize_db(self):
+        """Initialize the LMDB databases with data if they are empty."""
+        # Initialize users
+        with self.user_env.begin(write=True) as txn:
+            if not txn.stat()['entries']:
+                with txn.cursor() as cursor:
+                    for user in tqdm(self._iter_file('user.json')):
+                        cursor.put(
+                            user['user_id'].encode(),
+                            json.dumps(user).encode()
+                        )
+
+        # Initialize items
+        with self.item_env.begin(write=True) as txn:
+            if not txn.stat()['entries']:
+                with txn.cursor() as cursor:
+                    for item in tqdm(self._iter_file('item.json')):
+                        cursor.put(
+                            item['item_id'].encode(),
+                            json.dumps(item).encode()
+                        )
+
+        # Initialize reviews and their indices
+        with self.review_env.begin(write=True) as txn:
+            if not txn.stat()['entries']:
+                for review in tqdm(self._iter_file('review.json')):
+                    # Store the review
+                    txn.put(
+                        review['review_id'].encode(),
+                        json.dumps(review).encode()
+                    )
+
+                    # Update item reviews index (store only review_ids)
+                    item_review_ids = json.loads(txn.get(f"item_{review['item_id']}".encode()) or '[]')
+                    item_review_ids.append(review['review_id'])
+                    txn.put(
+                        f"item_{review['item_id']}".encode(),
+                        json.dumps(item_review_ids).encode()
+                    )
+
+                    # Update user reviews index (store only review_ids)
+                    user_review_ids = json.loads(txn.get(f"user_{review['user_id']}".encode()) or '[]')
+                    user_review_ids.append(review['review_id'])
+                    txn.put(
+                        f"user_{review['user_id']}".encode(),
+                        json.dumps(user_review_ids).encode()
+                    )
 
     def _iter_file(self, filename: str) -> Iterator[Dict]:
         """Iterate through file line by line."""
@@ -31,71 +86,55 @@ class CacheInteractionTool:
 
     def get_user(self, user_id: str) -> Optional[Dict]:
         """Fetch user data based on user_id."""
-        user = self.user_cache.get(user_id)
-        if user is not None:
-            return user
-        
-        for user in self._iter_file('user.json'):
-            if user['user_id'] == user_id:
-                self.user_cache[user_id] = user
-                return user
+        with self.user_env.begin() as txn:
+            user_data = txn.get(user_id.encode())
+            if user_data:
+                return json.loads(user_data)
         return None
 
     def get_item(self, item_id: str) -> Optional[Dict]:
         """Fetch item data based on item_id."""
         if not item_id:
             return None
-            
-        item = self.item_cache.get(item_id)
-        if item is not None:
-            return item
-        
-        for item in self._iter_file('item.json'):
-            if item['item_id'] == item_id:
-                self.item_cache[item_id] = item
-                return item
+
+        with self.item_env.begin() as txn:
+            item_data = txn.get(item_id.encode())
+            if item_data:
+                return json.loads(item_data)
         return None
 
     def get_reviews(
-        self, 
-        item_id: Optional[str] = None, 
-        user_id: Optional[str] = None, 
-        review_id: Optional[str] = None
+            self,
+            item_id: Optional[str] = None,
+            user_id: Optional[str] = None,
+            review_id: Optional[str] = None
     ) -> List[Dict]:
         """Fetch reviews filtered by various parameters."""
         if review_id:
-            review = self.review_cache.get(review_id)
-            if review is not None:
-                return [review]
-            
-            for review in self._iter_file('review.json'):
-                if review['review_id'] == review_id:
-                    self.review_cache[review_id] = review
-                    return [review]
+            with self.review_env.begin() as txn:
+                review_data = txn.get(review_id.encode())
+                if review_data:
+                    return [json.loads(review_data)]
             return []
 
-        if item_id:
-            cached_reviews = self.item_reviews_cache.get(item_id)
-            if cached_reviews is not None:
-                return cached_reviews
-            
+        with self.review_env.begin() as txn:
+            if item_id:
+                review_ids = json.loads(txn.get(f"item_{item_id}".encode()) or '[]')
+            elif user_id:
+                review_ids = json.loads(txn.get(f"user_{user_id}".encode()) or '[]')
+            else:
+                return []
+
+            # Fetch complete review data for each review_id
             reviews = []
-            for review in self._iter_file('review.json'):
-                if review['item_id'] == item_id:
-                    reviews.append(review)
-            self.item_reviews_cache[item_id] = reviews
+            for rid in review_ids:
+                review_data = txn.get(rid.encode())
+                if review_data:
+                    reviews.append(json.loads(review_data))
             return reviews
-            
-        elif user_id:
-            cached_reviews = self.user_reviews_cache.get(user_id)
-            if cached_reviews is not None:
-                return cached_reviews
-                
-            reviews = []
-            for review in self._iter_file('review.json'):
-                if review['user_id'] == user_id:
-                    reviews.append(review)
-            self.user_reviews_cache[user_id] = reviews
-            return reviews
-        
-        return []
+
+    def __del__(self):
+        """Cleanup LMDB environments on object destruction."""
+        self.user_env.close()
+        self.item_env.close()
+        self.review_env.close()