## The time for CacheInteractionTool is too long.
So I address the issues by implementing LMDB (Lightning Memory-Mapped Database) to significantly improve data retrieval performance. 
### 🔍 Problem
Previously, the system was experiencing severe performance limitations:
- ~1 second for getting a single item (get_item)
- ~50 seconds for getting a single history review (get_reviews)

### 💡 Solution
Implemented LMDB to optimize data storage and retrieval

### 📊 Performance Improvements

Before:
```
Time for getting item 0: 1.09 seconds
Time for getting item 1: 0.59 seconds
Time for getting item 2: 0.99 seconds
Time for getting item 3: 2.19 seconds
Time for getting item 4: 1.94 seconds
Time for getting item 5: 1.77 seconds
Time for getting item 6: 2.16 seconds
Time for getting item 7: 1.26 seconds
Time for getting item 8: 2.00 seconds
Time for getting item 9: 1.96 seconds
Time for getting item 10: 1.72 seconds
Time for getting item 11: 2.07 seconds
Time for getting item 12: 1.96 seconds
Time for getting item 13: 2.25 seconds
Time for getting item 14: 2.02 seconds
Time for getting item 15: 2.05 seconds
Time for getting item 16: 1.27 seconds
Time for getting item 17: 0.84 seconds
Time for getting item 18: 0.74 seconds
Time for getting item 19: 0.54 seconds
Time for history review: 49.82 seconds
```

After:
```
Database build time: ~10 minutes
Initializing InteractionTool with data directory: data/processed
889698it [00:14, 61311.70it/s]
358923it [00:23, 15160.01it/s]
5171890it [10:45, 8008.76it/s]
Time for getting user: 0.0026428699493408203
Time for getting item 0: 0.0008666515350341797
Time for getting item 1: 0.0009992122650146484
Time for getting item 2: 0.0010044574737548828
Time for getting item 3: 0.00099945068359375
Time for getting item 4: 0.0009999275207519531
Time for getting item 5: 0.0010001659393310547
Time for getting item 6: 0.0
Time for getting item 7: 0.0
Time for getting item 8: 0.0010006427764892578
Time for getting item 9: 0.00099945068359375
Time for getting item 10: 0.0005040168762207031
Time for getting item 11: 0.0010042190551757812
Time for getting item 12: 0.0
Time for getting item 13: 0.0010035037994384766
Time for getting item 14: 0.0010004043579101562
Time for getting item 15: 0.0
Time for getting item 16: 0.0010008811950683594
Time for getting item 17: 0.0
Time for getting item 18: 0.0
Time for getting item 19: 0.0010006427764892578
Time for getting history review: 0.0009987354278564453
```