@PinkGranite 2025-01-11
I think this is good. Do this support multi-thread access?