@PinkGranite 2025-01-11
During the official evaluation, we will not use CacheTool but directly load all data into memory for information retrieval, so this is not a problem in evaluation. However, for normal participants, your work may make great work so, If this has no issue with accessing data, especially if this is thread-safe, then I will approve it.